+++ pythonFiles/include/blender_vscode/installation.py
@@ -73,7 +73,10 @@ def get_package_install_directory() -> str:
 
 def module_can_be_imported(name: str):
     try:
-        __import__(_strip_pip_version(name))
+        stripped_name = _strip_pip_version(name)
+        mod = __import__(stripped_name)
+        LOG.info("module: " + name + " is already installed")
+        LOG.debug(stripped_name + ":" + getattr(mod ,"__version__", "None") + " in path: " + getattr(mod, "__file__", "None"))
         return True
     except ModuleNotFoundError:
         return False